+++ .flutter-plugins-dependencies
@@ -1,1 +1,1 @@
-{"info":"This is a generated file; do not edit or check into version control.","plugins":{"ios":[{"name":"flutter_email_sender","path":"C:\\\\Users\\\\cathy\\\\AppData\\\\Local\\\\Pub\\\\Cache\\\\hosted\\\\pub.dartlang.org\\\\flutter_email_sender-5.1.0\\\\","native_build":true,"dependencies":[]}],"android":[{"name":"flutter_email_sender","path":"C:\\\\Users\\\\cathy\\\\AppData\\\\Local\\\\Pub\\\\Cache\\\\hosted\\\\pub.dartlang.org\\\\flutter_email_sender-5.1.0\\\\","native_build":true,"dependencies":[]}],"macos":[],"linux":[],"windows":[],"web":[]},"dependencyGraph":[{"name":"flutter_email_sender","dependencies":[]}],"date_created":"2022-08-24 01:50:36.490167","version":"3.0.5"}
+{"info":"This is a generated file; do not edit or check into version control.","plugins":{"ios":[{"name":"flutter_email_sender","path":"C:\\\\Users\\\\cathy\\\\AppData\\\\Local\\\\Pub\\\\Cache\\\\hosted\\\\pub.dartlang.org\\\\flutter_email_sender-5.1.0\\\\","native_build":true,"dependencies":[]}],"android":[{"name":"flutter_email_sender","path":"C:\\\\Users\\\\cathy\\\\AppData\\\\Local\\\\Pub\\\\Cache\\\\hosted\\\\pub.dartlang.org\\\\flutter_email_sender-5.1.0\\\\","native_build":true,"dependencies":[]}],"macos":[],"linux":[],"windows":[],"web":[]},"dependencyGraph":[{"name":"flutter_email_sender","dependencies":[]}],"date_created":"2022-08-24 11:46:59.030682","version":"3.0.5"}
