+++ lib/main.dart
@@ -3,6 +3,7 @@ import 'package:calculator_frontend/resume_HoldingTax.dart';
 import 'package:calculator_frontend/widgets/HomePage/LargeLayout.dart';
 import 'package:calculator_frontend/widgets/HomePage/MediumLayout.dart';
 import 'package:flutter/material.dart';
+
 //2022월 08월 16일 22시 22분
 void main() {
   runApp(const MyApp());
