+++ pubspec.yaml
@@ -31,6 +31,7 @@ dependencies:
   csv: ^5.0.1
   http: ^0.13.4
   badges: ^2.0.3
+  http_client_helper: ^2.0.3
   flutter:
     sdk: flutter
